started to change the lib structure to group all base and user classes in the separate files
